update PR link checker workflow
Use a different changed-files checker (the old one was no longer maintained) and also make the link checking more fine-grained by only checking added lines for broken links. For this, we need to download lychee ourselves, because there is no Ubuntu package (there is a Conda one, but downloading a pre-built binary is a lot faster).